Fix metrics codegen when using label structs
Thanks Mason Stewart for the bug report.
