+++ compiler/internal/codegen/testdata/TestCodeGen_TestMain__metrics.golden
@@ -1,0 +1,35 @@
+// pkg svc
+package svc_test
+
+import (
+	"encore.app/svc"
+	__api "encore.dev/appruntime/api"
+	__appinit "encore.dev/appruntime/app/appinit"
+	__config "encore.dev/appruntime/config"
+	"os"
+	_ "unsafe"
+)
+
+//go:linkname loadApp encore.dev/appruntime/app/appinit.load
+func loadApp() *__appinit.LoadData {
+	os.Setenv("ENCORE_DUMMY_ENV_VAR", "eyAidGVzdCI6IHRydWUgfQ")
+	static := &__config.Static{
+		AuthData:             nil,
+		BundledServices:      []string{"svc"},
+		PubsubTopics:         map[string]*__config.StaticPubsubTopic{},
+		TestAsExternalBinary: true,
+		TestService:          "svc",
+		Testing:              true,
+	}
+	handlers := []__api.HandlerRegistration{
+		{
+			Handler:    svc.EncoreInternal_DummyHandler,
+			Middleware: nil,
+		},
+	}
+	return &__appinit.LoadData{
+		APIHandlers: handlers,
+		StaticCfg:   static,
+	}
+}
+
